src/scripts/requirements.txt
streamlit
psycopg2-binary
boto3
pgvector
